Refactor shelfMark stuff
